backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
psycopg2
python-dotenv
